Add "Force" parameter to ZipArchive.Constructor
When True, the Central Directory is ignored and the zip stream is assumed to start at offset zero. Checksums will be verified if ValifdateChecksums is True, however the MoveNext() method will not consider a checksum mismatch to be an error; it will set the last error code to ERR_CHECKSUM_MISMATCH, but Return True instead of False.