Add WriteEntryFooter method and OUTPUT_SEEKABLE constant
If OUTPUT_SEEKABLE is set to False, the ZipWriter class will append a desriptor after each file entry containing the crc32 and compressed size of the entry. This allows the compressor to avoid seeking backwards in the output stream to fill-in these values once they are known (i.e. *after* compression)