Add ZipReader.CurrentDataOffset
Returns the offset in the stream where the current item's compressed data begins. Combined with the CurrentSize and CurrentMethod properties you have all the information needed to decompress the raw file data. This will make it easier to build a table of contents that can be used for random access.